requirements.txt
mysql-connector-python==9.0.0
prettytable==3.10.0
PyYAML==6.0.1
wcwidth==0.2.13
